everything is in, style cleanup

+++ server_manager/www/app.ts
@@ -1102,7 +1102,7 @@ export class App {
       const tunnelTimeFomatter = Intl.NumberFormat(this.appRoot.language, {
         style: 'unit',
         unit: 'hour',
-        unitDisplay: 'short',
+        unitDisplay: 'long',
       });
 
       serverView.tunnelTimeTotal = tunnelTimeFomatter.format(
